adapt to solidity 0.7.0

+++ framework/src/test/resources/soliditycode/TvmOldCommand001.sol
@@ -1,11 +1,11 @@
 
 
 contract binaryRightContract{
-    function binaryMoveR(int i)public returns (int z) {
-        return z = 5 >> i;
+    function binaryMoveR(uint i)public returns (uint z) {
+        return z = uint(5) >> i;
     }
-    function binaryLiftR(int i)public returns (int z) {
-        return z = 5 << i;
+    function binaryLiftR(uint i)public returns (uint z) {
+        return z = uint(5) << i;
     }
 
 }
